Fix setNear/FarDistance with proj recalculus

+++ GUIInspector.cpp
@@ -32,8 +32,8 @@ void GUIInspector::Draw(const char * title, bool * p_opened)
 		ImGui::TextColored(ImVec4(1.0f, 0.0, 1.0f, 1.0f), "Current textures loaded");
 		for(auto index : App->modelLoader->indicesOfCurrentTextures)
 		{
-			int window_width = ImGui::GetWindowWidth();
-			int window_height = ImGui::GetWindowHeight();
+			int window_width = (int) ImGui::GetWindowWidth();
+			int window_height = (int) ImGui::GetWindowHeight();
 			ImGui::TextColored(ImVec4(1.0f, 1.0f, 0.0f, 1.0f), "Texture path: "); ImGui::SameLine();
 			ImGui::Text("%s", App->texture->textures_loaded[index].path.c_str());
 			ImGui::Image((ImTextureID)App->texture->textures_loaded[index].id, ImVec2(window_height * 0.5f, window_width * 0.5f), ImVec2(0, 1), ImVec2(1, 0));
