+++ ModuleInput.cpp
@@ -122,8 +122,8 @@ update_status ModuleInput::PreUpdate()
 			break;
 
 		case SDL_MOUSEMOTION:
-			mouse_motion.x = event.motion.xrel;
-			mouse_motion.y = event.motion.yrel;
+			mouse_motion.x = (float)event.motion.xrel;
+			mouse_motion.y = (float)event.motion.yrel;
 			mouse.x = (float)event.motion.x / SCREEN_WIDTH;
 			mouse.y = (float)event.motion.y / SCREEN_HEIGHT;
